Increased efficiency of vertex transformations
Added the ability to reset the position and rotation
of 3d objects.

Added command to object class which allows the user
to explicitly set the rotation of the object instead
of a rotation relative to its current position.

Changed the vertices of the triangle class to
pointers. Triangles now only store pointers to
their vertices instead of copying them. The object
class now only transforms the positions of its own
unique vertices instead of transforming the vertices
of all triangles it contains. This should greatly
increase efficiency for more complex objects.