Implemented camera rotation
Added rotation methods to camera class as well as
the ability to point the camera at a point in 3d
space.

Added position methods to camera class as well as
methods which move the camera relative to its
current position.

Fixed camera class point projection mathematical
error which caused the projection on the viewing
plane to be incorrect.

Camera class now longer inherits from ray class.