Added polygon render screen clipping
Added screen clipping of rendered polygons which
prevents polygons being rendered off the viewable
screen.

Added a clipping border which may be used to offset
the clipping boundary away from the physical extents
of the screen (mostly for testing).

Fixed rendering error which caused polygons behind
the camera to be rendered.