Updated renderer executable to include recent updates
Updated the complexity of the default scene in the renderer test executable to take recent updates and improvements into account.

The scene class no longer automatically renders the scene after drawing objects to the frame buffer. This allows further processing to be done after objects are drawn.

Added cmake variable "ASSETS_DIRECTORY" which may be used to specify a local assets path to avoid hard-coded filenames.